Set FlipView's dataSource in ready(), make itemTemplate conf more robust
The item object is only available in ready() since it's passed to
item control when navigating to it. Therefore, it cannot be used as
a dataSource when creating the FlipView, the dataSource needs to
set to FlipView later on.

Alternative approach would be to have a static, empty dataSource in item
control, set that declaratively to be FlipView's dataSource and then fill
it up in ready() once the item object is available.